bot/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
openai==1.14.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
redis==5.0.3
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
yarl==1.9.4
